+++ packages/nimble/inst/include/nimble/nimDerivs_atomic_solve_base.h
@@ -28,7 +28,7 @@ class atomic_solve_base_class {
   CppAD::AD<double> * get_X_AD_stored_ptr() {return &X_AD_stored[0];}
   void fill_X_AD_stored() {
     X_AD_stored.resize( X_stored.size() );
-    for(int i = 0; i < X_stored.size(); ++i) X_AD_stored[i] = X_stored[i];
+    for(size_t i = 0; i < X_stored.size(); ++i) X_AD_stored[i] = X_stored[i];
   };
   void clear_X_AD_stored() {X_AD_stored.clear();}
   std::vector<double> &get_X_stored() {return X_stored;}
